:bug: remove conditional before hook
end up not working well on unmount :shrug:
